+++ clients/roscpp/include/ros/service_publication.h
@@ -30,7 +30,7 @@
 
 #include "ros/service_callback_helper.h"
 #include "common.h"
-#include "XmlRpc.h"
+#include "xmlrpcpp/XmlRpc.h"
 
 #include <boost/thread/mutex.hpp>
