+++ clients/roscpp/src/libros/service_manager.cpp
@@ -42,7 +42,7 @@
 #include "ros/connection.h"
 #include "ros/file_log.h"
 
-#include "XmlRpc.h"
+#include "xmlrpcpp/XmlRpc.h"
 
 #include <ros/console.h>
