+++ utilities/xmlrpcpp/cmake/xmlrpcpp-extras.cmake.in
@@ -1,0 +1,1 @@
+list(APPEND @PROJECT_NAME@_INCLUDE_DIRS "${@PROJECT_NAME@_DIR}/../../../@CATKIN_GLOBAL_INCLUDE_DESTINATION@/xmlrpcpp")
