+++ utilities/xmlrpcpp/include/xmlrpcpp/XmlRpcException.h
@@ -12,7 +12,7 @@
 # include <string>
 #endif
 
-#include "XmlRpcDecl.h"
+#include "xmlrpcpp/XmlRpcDecl.h"
 
 
 namespace XmlRpc {
