+++ utilities/xmlrpcpp/src/XmlRpcClient.cpp
@@ -1,8 +1,8 @@
 
-#include "XmlRpcClient.h"
+#include "xmlrpcpp/XmlRpcClient.h"
 
-#include "XmlRpcSocket.h"
-#include "XmlRpc.h"
+#include "xmlrpcpp/XmlRpcSocket.h"
+#include "xmlrpcpp/XmlRpc.h"
 
 #include <stdio.h>
 #include <stdlib.h>
