+++ utilities/xmlrpcpp/test/FileClient.cpp
@@ -3,7 +3,7 @@
 // 
 // Link against xmlrpc lib and whatever socket libs your system needs (ws2_32.lib on windows)
 
-#include "XmlRpc.h"
+#include "xmlrpcpp/XmlRpc.h"
 #include <iostream>
 #include <fstream>
 #include <stdlib.h>
