+++ utilities/xmlrpcpp/test/HelloClient.cpp
@@ -1,7 +1,7 @@
 // HelloClient.cpp : A simple xmlrpc client. Usage: HelloClient serverHost serverPort
 // Link against xmlrpc lib and whatever socket libs your system needs (ws2_32.lib 
 // on windows)
-#include "XmlRpc.h"
+#include "xmlrpcpp/XmlRpc.h"
 #include <iostream>
 using namespace XmlRpc;
