+++ utilities/xmlrpcpp/test/TestValuesWin32.cpp
@@ -4,7 +4,7 @@
 #include <stdlib.h>
 #include <crtdbg.h>
 
-#include "XmlRpcValue.h"
+#include "xmlrpcpp/XmlRpcValue.h"
 
 
 #include <assert.h>
